feat: set context for search tearn
